pyproject.toml
[tool.poetry]
name = "gas-sensors-service"
version = "0.1.0"
description = ""
authors = ["Andrey Gaev <a.v.gaev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
scikit-learn = "^1.4.2"
tqdm = "^4.66.4"
seaborn = "^0.13.2"
pyyaml = "^6.0.1"
flask = "^3.0.3"
redis = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
